copy updated files from gitee to github

+++ cplusplus/contrib/ar_shadowgan/scripts/sample_build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
-. ${ScriptPath}/../../../../common/sample_common.sh
+common_script_dir=${THIRDPART_PATH}/common
+. ${common_script_dir}/sample_common.sh
 
 function main()
 {
@@ -11,10 +12,12 @@ function main()
   if [ $? -ne 0 ];then
     return 1
   fi
-
-  wget -O ${ModelPath}/../data/test_A.png.bin https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/arshadow/data/test_A.png.bin --no-check-certificate
-  wget -O ${ModelPath}/../data/test_B.png.bin https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/arshadow/data/test_B.png.bin --no-check-certificate
-  
+  if [ ! -f "${ModelPath}/../data/test_A.png.bin" ];then
+    wget -O ${ModelPath}/../data/test_A.png.bin https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/arshadow/data/test_A.png.bin --no-check-certificate
+  fi
+  if [ ! -f "${ModelPath}/../data/test_B.png.bin" ];then
+    wget -O ${ModelPath}/../data/test_B.png.bin https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/arshadow/data/test_B.png.bin --no-check-certificate
+  fi
   find_model model.om
   if [ $? -ne 0 ];then
     return 1
