+++ cplusplus/contrib/cartoonGAN_picture/scripts/sample_build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
-. ${ScriptPath}/../../../../common/sample_common.sh
+common_script_dir=${THIRDPART_PATH}/common
+. ${common_script_dir}/sample_common.sh
 
 function main()
 {
@@ -11,9 +12,9 @@ function main()
   if [ $? -ne 0 ];then
     return 1
   fi
-
-  wget -O ${ModelPath}/../data/test.jpg https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/cartoonGAN_picture/test_image/test.jpg --no-check-certificate
-  
+  if [ ! -f "${ModelPath}/../data/test.jpg" ];then
+    wget -O ${ModelPath}/../data/test.jpg https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/cartoonGAN_picture/test_image/test.jpg --no-check-certificate
+  fi
   find_model cartoonization.om
   if [ $? -ne 0 ];then
     return 1
