+++ cplusplus/contrib/gesturedetection/scripts/sample_build.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
-. ${ScriptPath}/../../../../common/sample_common.sh
+common_script_dir=${THIRDPART_PATH}/common
+. ${common_script_dir}/sample_common.sh
 
 data_source="https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/gesturedetection/test_image/"
 project_name="cplusplus_Gesturedetection"
 
 project_path=${ScriptPath}/..
-
+jpg=".jpg"
 function main()
 {
   echo "[INFO] Sample preparation"
@@ -20,14 +21,14 @@ function main()
 
     for ((i=0; i<50; i ++))
     do
-    wget -O ${project_path}/data/$i".jpg"  ${data_source}$i".jpg"  --no-check-certificate
+    if [ ! -f "${ModelPath}/../data/${i}${jpg}" ];then
+        wget -O ${project_path}/data/$i".jpg"  ${data_source}$i".jpg"  --no-check-certificate
+    fi
     if [ $? -ne 0 ];then
             echo "Download jpg failed, please check network."
             return 1
         fi
     done
-
- # wget -O ${ModelPath}/../data/0.jpg wget https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/gesturedetection/test_image/0.jpg --no-check-certificate
   
   find_model pose_deploy.om
   find_model stgcn_fps30_sta_ho_ki4.om
