+++ cplusplus/contrib/human_segmentation/scripts/sample_build.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
-common_script_dir=${ScriptPath}/../../../../common/
+common_script_dir=${THIRDPART_PATH}/common
 . ${common_script_dir}/sample_common.sh
 
 function main()
@@ -12,9 +12,9 @@ function main()
   if [ $? -ne 0 ];then
     return 1
   fi
-
-  wget -O ${ModelPath}/../data/person.mp4 https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/human_segmentation/person.mp4 --no-check-certificate
-  
+  if [ ! -f "${ModelPath}/../data/person.mp4" ];then
+    wget -O ${ModelPath}/../data/person.mp4 https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/human_segmentation/person.mp4 --no-check-certificate
+  fi
   find_model human_segmentation.om
   if [ $? -ne 0 ];then
     return 1
