+++ cplusplus/contrib/super_resolution/scripts/sample_build.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
-. ${ScriptPath}/../../../../common/sample_common.sh
-
+common_script_dir=${THIRDPART_PATH}/common
+. ${common_script_dir}/sample_common.sh
+echo ${common_script_dir}
 function main()
 {
   echo "[INFO] Sample preparation"
@@ -11,8 +12,9 @@ function main()
   if [ $? -ne 0 ];then
     return 1
   fi
-
-  wget -O ${ModelPath}/../data/butterfly_GT.bmp https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/super_resolution/test_image/butterfly_GT.bmp --no-check-certificate
+  if [ ! -f "${ModelPath}/../data/butterfly_GT.bmp" ];then
+    wget -O ${ModelPath}/../data/butterfly_GT.bmp https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/super_resolution/test_image/butterfly_GT.bmp --no-check-certificate
+  fi
   
   find_model SRCNN_768_768.om
   if [ $? -ne 0 ];then
