+++ cplusplus/level2_simple_inference/2_object_detection/YOLOV3_coco_detection_picture_with_freetype/scripts/sample_build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
-. ${ScriptPath}/../../../../../common/sample_common.sh
+common_script_dir=${THIRDPART_PATH}/common
+. ${common_script_dir}/sample_common.sh
 
 function main()
 {
@@ -12,7 +13,9 @@ function main()
     return 1
   fi
 
-  wget -O ${ModelPath}/../data/dog_1024_688.jpg https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/YOLOV3_coco_detection_picture_with_freetype/dog_1024_688.jpg --no-check-certificate
+  if [ ! -f "${ModelPath}/../data/dog_1024_688.jpg" ];then
+    wget -O ${ModelPath}/../data/dog_1024_688.jpg https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/YOLOV3_coco_detection_picture_with_freetype/dog_1024_688.jpg --no-check-certificate
+  fi
   
   find_model yolov3.om
   if [ $? -ne 0 ];then
@@ -27,5 +30,4 @@ function main()
   echo "[INFO] Sample preparation is complete"
 }
 main
-
