+++ cplusplus/level2_simple_inference/2_object_detection/YOLOV3_coco_detection_video/scripts/sample_build.sh
@@ -2,7 +2,8 @@
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
 
-. ${ScriptPath}/../../../../../common/sample_common.sh
+common_script_dir=${THIRDPART_PATH}/common
+. ${common_script_dir}/sample_common.sh
 
 function main()
 {
@@ -13,8 +14,10 @@ function main()
     return 1
   fi
 
-  wget -O ${ModelPath}/../data/detection.mp4 https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/YOLOV3_coco_detection_video/detection.mp4 --no-check-certificate
-  
+  if [ ! -f "${ModelPath}/../data/detection.mp4" ];then
+    wget -O ${ModelPath}/../data/detection.mp4 https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/YOLOV3_coco_detection_video/detection.mp4 --no-check-certificate
+  fi
+
   find_model yolov3.om
   if [ $? -ne 0 ];then
     return 1
