+++ cplusplus/level2_simple_inference/2_object_detection/YOLOV4_coco_detection_picture/scripts/sample_build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
-. ${ScriptPath}/../../../../../common/sample_common.sh
+common_script_dir=${THIRDPART_PATH}/common
+. ${common_script_dir}/sample_common.sh
 
 function main()
 {
@@ -12,8 +13,10 @@ function main()
     return 1
   fi
 
-  wget -O ${ModelPath}/../data/dog1_1024_683.jpg https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/aclsample/dog1_1024_683.jpg --no-check-certificate
-  
+  if [ ! -f "${ModelPath}/../data/dog1_1024_683.jpg" ];then
+    wget -O ${ModelPath}/../data/dog1_1024_683.jpg https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/aclsample/dog1_1024_683.jpg --no-check-certificate
+  fi
+
   find_model yolov4.om
   if [ $? -ne 0 ];then
     return 1
