+++ cplusplus/level2_simple_inference/6_other/super_resolution_dynamic/scripts/sample_build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
-. ${ScriptPath}/../../../../../common/sample_common.sh
+common_script_dir=${THIRDPART_PATH}/common
+. ${common_script_dir}/sample_common.sh
 
 function main()
 {
@@ -12,8 +13,10 @@ function main()
     return 1
   fi
 
+  if [ ! -f "${ModelPath}/../data/butterfly_GT.bmp" ];then
   wget -O ${ModelPath}/../data/butterfly_GT.bmp https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/super_resolution/test_image/butterfly_GT.bmp --no-check-certificate
-  
+  fi
+
   find_model FSRCNN_x3.om
   if [ $? -ne 0 ];then
     return 1
