+++ cplusplus/level2_simple_inference/n_performance/1_multi_process_thread/multi_channels_rtsp/scripts/sample_build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 ScriptPath="$( cd "$(dirname "$BASH_SOURCE")" ; pwd -P )"
 ModelPath="${ScriptPath}/../model"
-. ${ScriptPath}/../../../../../../common/sample_common.sh
+common_script_dir=${THIRDPART_PATH}/common
+. ${common_script_dir}/sample_common.sh
 
 function main()
 {
@@ -12,7 +13,9 @@ function main()
     return 1
   fi
 
+  if [ ! -f "${ModelPath}/../data/person.mp4" ];then
   wget -O ${ModelPath}/../data/person.mp4 https://c7xcode.obs.cn-north-4.myhuaweicloud.com/models/face_detection_rtsp/person.mp4 --no-check-certificate
+  fi
   
   find_model face_detection.om
   if [ $? -ne 0 ];then
